third_party/WebKit/LayoutTests/inspector/styles/hidden-user-agent-styles-in-computed-trace.html

<html>
<head>

<style>
div {
    display: inline-block;
}
</style>

<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>

function test()
{
    WebInspector.settings.showUserAgentStyles.set(false);
    InspectorTest.addResult("WebInspector.settings.showUserAgentStyles = " + WebInspector.settings.showUserAgentStyles.get());
    InspectorTest.selectNodeAndWaitForStylesWithComputed("inspected", callback);

    function callback()
    {
        InspectorTest.dumpSelectedElementStyles();
        WebInspector.settings.showUserAgentStyles.set(true);
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<div id="inspected" />
<p>
Tests that user agent styles are shown in the computed trace regardless of the "Show user agent styles" setting: <a href="http://crbug.com/276293">http://crbug.com/276293</a>.
</p>
</body>
</html>
