third_party/WebKit/LayoutTests/inspector/styles/hidden-user-agent-styles-in-computed-trace-expected.txt

Tests that user agent styles are shown in the computed trace regardless of the "Show user agent styles" setting: http://crbug.com/276293.

WebInspector.settings.showUserAgentStyles = false
[expanded] 
display: inline-block;
    div - inline-block hidden-user-age…d-trace.html:5
    /-- overloaded --/ div - block user agent stylesheet

[expanded] 
element.style  { ()

[expanded] 
div  { (hidden-user-age…d-trace.html:5 -> hidden-user-agent-styles-in-computed-trace.html:5:1)
display: inline-block;
